cookie response & db session storage set up, bug fix
fixed bug where an invalid email login lookup returns status code 500 instead of a 401 (email not found in database)

implemented logic through which a successful login returns a cookie with a randomized id, and should insert it into the database provided it has the correct table and schema.

We don't actually do anything with the sessions yet, still have to implement use cases for that.